+++ src/app/useChainIdRoute.ts
@@ -1,4 +1,4 @@
-import { CHAIN_ID, isSupportedChainId } from '@/stores/blockchain/blockchain'
+import { CHAIN_ID, isSupportedChainId } from '@/stores/blockchain/chains'
 import { DEFAULT_CHAIN_ID, useBlockchain } from '@/stores/blockchain/useBlockchain'
 import { useRouter } from 'vue-router'
