+++ src/components/ImportAccountModal/ConfirmImport.vue
@@ -3,7 +3,7 @@ import { AccountRegistry } from '@/lib/accounts'
 import { ValidationMethodData } from '@/lib/validations/ValidationMethod'
 import { ImportedAccount } from '@/stores/account/account'
 import { useAccounts } from '@/stores/account/useAccounts'
-import { CHAIN_NAME } from '@/stores/blockchain/blockchain'
+import { CHAIN_NAME } from '@/stores/blockchain/chains'
 import { useBlockchain } from '@/stores/blockchain/useBlockchain'
 import { useImportAccountModal } from '@/stores/useImportAccountModal'
 import { shortenAddress } from '@vue-dapp/core'
